+++ tikv-client/src/main/java/com/pingcap/tikv/TiConfiguration.java
@@ -46,7 +46,7 @@ public class TiConfiguration implements Serializable {
   private static final int DEF_TABLE_SCAN_CONCURRENCY = 512;
   private static final CommandPri DEF_COMMAND_PRIORITY = CommandPri.Low;
   private static final IsolationLevel DEF_ISOLATION_LEVEL = IsolationLevel.RC;
-  private static final long REGION_INDEX_SCAN_DOWNGRADE_THRESHOLD = 100000;
+  private static final int REGION_INDEX_SCAN_DOWNGRADE_THRESHOLD = 100000;
 
   private int retryTimes = DEF_RETRY_TIMES;
   private int timeout = DEF_TIMEOUT;
@@ -63,7 +63,7 @@ public class TiConfiguration implements Serializable {
   private int tableScanConcurrency = DEF_TABLE_SCAN_CONCURRENCY;
   private CommandPri commandPriority = DEF_COMMAND_PRIORITY;
   private IsolationLevel isolationLevel = DEF_ISOLATION_LEVEL;
-  private long regionIndexScanDowngradeThreshold = REGION_INDEX_SCAN_DOWNGRADE_THRESHOLD;
+  private int regionIndexScanDowngradeThreshold = REGION_INDEX_SCAN_DOWNGRADE_THRESHOLD;
   private int maxRequestKeyRangeSize = MAX_REQUEST_KEY_RANGE_SIZE;
 
   public static TiConfiguration createDefault(String pdAddrsStr) {
@@ -218,11 +218,11 @@ public class TiConfiguration implements Serializable {
     this.isolationLevel = isolationLevel;
   }
 
-  public long getRegionIndexScanDowngradeThreshold() {
+  public int getRegionIndexScanDowngradeThreshold() {
     return regionIndexScanDowngradeThreshold;
   }
 
-  public void setRegionIndexScanDowngradeThreshold(long regionIndexScanDowngradeThreshold) {
+  public void setRegionIndexScanDowngradeThreshold(int regionIndexScanDowngradeThreshold) {
     this.regionIndexScanDowngradeThreshold = regionIndexScanDowngradeThreshold;
   }
